cart_service/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4
psycopg2-binary>=2.9,<3
gunicorn>=22.0,<23
requests>=2.32,<3
